+++ src/runtime/composables/useEchoAppConfig.ts
@@ -1,5 +1,5 @@
 import type { EchoAppConfig } from '../types/config'
-import { useAppConfig } from '#app'
+import { useAppConfig } from '#imports'
 
 export const useEchoAppConfig = (): EchoAppConfig => {
   return (useAppConfig().echo ?? {}) as EchoAppConfig